backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
pandas
scikit-learn
xgboost
networkx
python-dotenv
